feat: Add API functions
The function clear_teams was added. This function cleans the
teams global dict

The function get_new_id was added. This function returns a new
id that is not been used by another team

The function get_all_teams was added. This function returns all the
current teams

The function get_user_teams was added. This function returns a
team that belongs to the user passed as a parameter

The function create_team was added. This function add a team to
the teams global dict and returns a object with a successeful
message and the id of the new team